refactor: extract cosmos utils
